flask_backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
pymongo==4.0.1
bcrypt==3.2.0
python-dotenv==0.19.1
marshmallow==3.14.1
email-validator==1.1.3
flask-mongoengine==1.0.0
gunicorn==20.1.0
